Mv test utils to own file
